refactor: move useRedeemClaimTools to useDei hooks file
